@@ -1,9 +1,9 @@
-name: bhaaratcuisines
-description: A new Flutter application.
+name: bharatcuisines
+description: Wake up your taste buds.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+# publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=2.17.1 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,29 +29,35 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-
+  grouped_list: ^5.1.2
+  url_launcher: ^6.1.4
+  google_fonts: ^3.0.1
+  animated_splash_screen: ^1.2.0
+  flutter_native_splash: ^2.2.3+1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
-  grouped_list: ^5.1.2
-  url_launcher: ^6.1.4
+  
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
   flutter_lints: ^1.0.0
+  flutter_launcher_icons: ^0.9.3
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
+flutter_icons:
+  android: true
+  image_path: "assets/logo.png"
 
-# The following section is specific to Flutter.
+flutter_native_splash:
+  # color: "#42a5f5"
+  image: assets/background.png
+  android: true
+  android_gravity: center
+  fullscreen: true
+  android12: true
+
+
 flutter:
 
   # The following line ensures that the Material Icons font is
@@ -60,12 +66,14 @@ flutter:
   uses-material-design: true
 
   # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+    - assets/bg1.jpg
+    - assets/bg8.png
+    - assets/background.png
+   
 
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
+  # https://flutter.dev/assets-and-images/#resolution-aware
 
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
