@@ -1,17 +1,466 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_application_1/state_view.dart';
+// import 'package:flutter/material.dart';
 
-import './models/state.dart';
+// List _data = [
+//   {
+//     'id': 's2',
+//     'title': 'Madhya Pradesh',
+//     'color': Colors.purple,
+//     'group': 'Central India',
+//     'img':
+//         'https://img.traveltriangle.com/blog/wp-content/uploads/2019/05/Cover-for-Places-to-visit-in-Madhya-Pradesh-in-Summer.jpg',
+//     'cuisines': [
+//       {
+//         'cName': 'poha',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Daal Bafla',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Malpua',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Mohanthal',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//     ],
+//   },
+//   {
+//     'id': 's2',
+//     'title': 'Chhattisgarh',
+//     'color': Colors.purple,
+//     'group': 'Central India',
+//     'img':
+//         'https://www.nationsonline.org/gallery/India/Bhoramdeo-Temple-Kawardha.jpg',
+//     'cuisines': [
+//       {
+//         'cName': 'Muthia',
+//         'image':
+//             'https://masalachilli.com/wp-content/uploads/2020/12/Dudhi-Muthiya-4.jpg',
+//         'ingredients': [
+//           '2 cups besan (gram flour)',
+//           '2.5 cups finely chopped methi leaves',
+//           '1 teaspoon salt',
+//           '2 teaspoons sugar',
+//           '2 tablespoons fine rava or sooji (cream of wheat)',
+//           '2 teaspoons white sesame seeds',
+//           '1 teaspoon turmeric powder, coriander powder and coriander powder',
+//           '½ teaspoon red chili powder'
+//         ],
+//         'link': 'https://www.youtube.com/watch?v=Ft_2LIcvOU8',
+//         'recipe':
+//             '1. Combine to mix the ingredients thoroughly.\n2. Add 2 tablespoons ginger-green chili paste, ¼ teaspoon baking soda and 2 tablespoons lemon juice.\n'
+//                 'Now, add 1 tablespoon water. Remember to add water as needed and knead to a smooth dough.'
+//                 '                       '
+//                 '''
+// To make steamed muthia shape the dough into a length of 4 to 5 inches sausage type rolls.
+// Place them on a greased tray. Grease a baking tray or pan with neutral flavored oil. You can spread some oil in your palms while shaping the dough.
+// Grease a baking tray or pan with neutral flavored oil. You can spread some oil in your palms while shaping the dough.
+// Heat sufficient water or about 2 cups of water in a deep pan with a trivet placed on it. When the water becomes hot, using tongs carefully place the greased tray on the trivet in the pan. 
+// Cover and steam the muthiya for 10 to 12 minutes.
+// The muthia should be perfectly steamed and cooked. Check with a tooth pick or knife to see for doneness. 
+// If its properly steamed, then the dough will not stick to the tooth pick and will come out clean.
 
-const DUMMY_STATES = const [
-  StateName(
-    id: 'c1',
-    title: 'Italian',
-    color: Colors.purple,
-  ),
-  StateName(
-    id: 'c2',
-    title: 'Quick & Easy',
-    color: Colors.red,
-  ),
-];
+//  '''
+//                 '',
+//         'note':
+//             'Serve as a snack with ketchup or coriander chutney. It tastes best when served with mango chunda.'
+//       },
+//       {
+//         'cName': 'Bara',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups besan (gram flour)',
+//           '2.5 cups finely chopped methi leaves',
+//           '1 teaspoon salt',
+//           '2 teaspoons sugar',
+//           '2 tablespoons fine rava or sooji (cream of wheat)',
+//           '2 teaspoons white sesame seeds',
+//           '1 teaspoon turmeric powder, coriander powder and coriander powder',
+//           '½ teaspoon red chili powder'
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe': '',
+//         'note':
+//             ' Care should be taken to maintain the temperature at a constant 140°C.'
+//       },
+//       {
+//         'cName': 'Dal Pithi',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Arsa',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups besan (gram flour)',
+//           '2.5 cups finely chopped methi leaves',
+//           '1 teaspoon salt',
+//           '2 teaspoons sugar',
+//           '2 tablespoons fine rava or sooji (cream of wheat)',
+//           '2 teaspoons white sesame seeds',
+//           '1 teaspoon turmeric powder, coriander powder and coriander powder',
+//           '½ teaspoon red chili powder'
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//     ],
+//   },
+//   {
+//     'id': 's1',
+//     'title': 'Rajasthan',
+//     'color': Colors.amber,
+//     'group': 'Western Indian',
+//     'img':
+//         'https://dynamic-media-cdn.tripadvisor.com/media/photo-o/15/33/fc/f9/rajasthan.jpg?w=700&h=500&s=1',
+//     'cuisines': [
+//       {
+//         'cName': 'Dal Bati Churma',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Pyaz ki Kachori',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Gatte ki Sabzi',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Mohanthal',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//     ],
+//   },
+//   {
+//     'id': 's1',
+//     'title': 'Maharashtra',
+//     'color': Colors.amber,
+//     'group': 'Western Indian',
+//     'img':
+//         'https://media.istockphoto.com/photos/gateway-of-india-mumbai-maharashtra-monument-landmark-famous-place-picture-id1307189136?b=1&k=20&m=1307189136&s=170667a&w=0&h=F-BSrbduokK7pz1AEOwOGuoDe2xd28wnSyZpll4Jzks=',
+//     'cuisines': [
+//       {
+//         'cName': 'Dal Bati Churma',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Pyaz ki Kachori',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Gatte ki Sabzi',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Mohanthal',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//     ],
+//   },
+//   {
+//     'id': 's1',
+//     'title': 'Gujarat',
+//     'color': Colors.amber,
+//     'group': 'Western Indian',
+//     'img':
+//         'https://static.toiimg.com/photo/msid-84602392,width-96,height-65.cms',
+//     'cuisines': [
+//       {
+//         'cName': 'Dal Bati Churma',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Pyaz ki Kachori',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Gatte ki Sabzi',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//       {
+//         'cName': 'Mohanthal',
+//         'image':
+//             'https://media.vogue.in/wp-content/uploads/2020/10/poha-recipe-1920x1080.jpg',
+//         'ingredients': [
+//           '2 cups thick poha (flattened rice)',
+//           '1 medium onion finely, chopped',
+//           '1 medium potato, chopped into small cubes(1/4 inch cube)',
+//           '1/4 cup frozen green peas, or veggies of choice; optional',
+//           '',
+//           '',
+//           '',
+//           ''
+//         ],
+//         'link': 'https://youtu.be/XMrGrd011wQ',
+//         'recipe':
+//             'To a sieve, add 2 cups of poha, and wash it under running water for a few minutes. Place the sieve in a bowl and let the water drain completely, around 10-15 minutes. This helps soften the poha.',
+//         'note':
+//             'Check poha if it is soft enough. Take a poha flake and break it between the fingers. It should easily crush.'
+//       },
+//     ],
+//   },
+// ];
